add reverb

+++ src/denoiser/modeling/waveunet/model.py
@@ -154,7 +154,7 @@ class WaveUNet(pl.LightningModule):
 
         o = torch.cat([o, inputs], dim=1)
         o = self.out(o)
-        return o
+        return o.to(torch.float32)
 
     def training_step(
         self, batch: Sample, batch_idx: Any
